temporarily resize for debugging and balance start to unit stats
All panels are visible at once and it uses entire screen so it is easy to make sure everything works. It is easy to change this back. Marine is now Scout and stats are sort of balanced. Double click a unit to toggle attack/defense. Health bar color changes depending on attack/defense status. We can now play actually play the game to test for balance (except for AirStrike).